refactor(web): add annotation queue menu controller
