+++ web/src/components/session/SessionTraceActionButtons.tsx
@@ -4,7 +4,9 @@ import { getNumberFromMap } from "@/src/utils/map-utils";
 import { AnnotateDrawer } from "@/src/features/scores/components/AnnotateDrawer";
 import { CommentDrawerButton } from "@/src/features/comments/CommentDrawerButton";
 import { NewDatasetItemFromTraceId } from "@/src/components/session/NewDatasetItemFromTrace";
-import { CreateNewAnnotationQueueItem } from "@/src/features/annotation-queues/components/CreateNewAnnotationQueueItem";
+import { Button } from "@/src/components/ui/button";
+import { AnnotationQueueItemDropdownMenuController } from "@/src/features/annotation-queues/components/AnnotationQueueItemDropdownMenuController";
+import { ChevronDown } from "lucide-react";
 import { cn } from "@/src/utils/tailwind";
 
 type TraceScores =
@@ -60,13 +62,29 @@ export function SessionTraceActionButtons({
             environment: environment ?? undefined,
           }}
         />
-        <CreateNewAnnotationQueueItem
+        <AnnotationQueueItemDropdownMenuController
           projectId={projectId}
           objectId={traceId}
           objectType="TRACE"
-          variant="outline"
-          size={size}
-        />
+        >
+          {({ disabled, totalCount }) => (
+            <Button
+              variant="outline"
+              size={size}
+              disabled={disabled !== undefined}
+              className="rounded-l-none rounded-r-md border-l-2"
+            >
+              <span className="relative mr-1 text-xs">
+                <ChevronDown className="h-3 w-3" />
+                {totalCount > 0 ? (
+                  <span className="bg-primary text-primary-foreground absolute -top-1 left-2.5 flex h-3 min-w-3 items-center justify-center rounded-sm px-0.5 text-[8px] font-bold shadow-xs">
+                    {totalCount > 99 ? "99+" : totalCount}
+                  </span>
+                ) : null}
+              </span>
+            </Button>
+          )}
+        </AnnotationQueueItemDropdownMenuController>
       </div>
       <CommentDrawerButton
         projectId={projectId}
