+++ web/src/features/traces/components/TraceDetailView/components/TraceDetailViewHeader.tsx
@@ -24,7 +24,7 @@ import { LocalIsoDate } from "@/src/components/LocalIsoDate";
 import { DetailHeaderActionsMenu } from "@/src/features/traces/components/DetailHeaderActionsMenu";
 import { NewDatasetItemFromExistingObject } from "@/src/features/datasets/components/NewDatasetItemFromExistingObject";
 import { AnnotateDrawer } from "@/src/features/scores/components/AnnotateDrawer";
-import { CreateNewAnnotationQueueItem } from "@/src/features/annotation-queues/components/CreateNewAnnotationQueueItem";
+import { AnnotationQueueItemDropdownMenuController } from "@/src/features/annotation-queues/components/AnnotationQueueItemDropdownMenuController";
 import { CommentDrawerButton } from "@/src/features/comments/CommentDrawerButton";
 import {
   SessionBadge,
@@ -42,7 +42,7 @@ import { useViewPreferences } from "@/src/features/traces/contexts/ViewPreferenc
 import { CollapsibleBadgeRow } from "@/src/features/traces/components/CollapsibleBadgeRow";
 import { useIsMobile } from "@/src/hooks/use-mobile";
 import { Button } from "@/src/components/ui/button";
-import { MoreHorizontal } from "lucide-react";
+import { ChevronDown, ListPlus, MoreHorizontal } from "lucide-react";
 import {
   Popover,
   PopoverContent,
@@ -161,12 +161,28 @@ export const TraceDetailViewHeader = memo(function TraceDetailViewHeader({
                       }}
                       layout="menu"
                     />
-                    <CreateNewAnnotationQueueItem
+                    <AnnotationQueueItemDropdownMenuController
                       projectId={projectId}
                       objectId={trace.id}
                       objectType={AnnotationQueueObjectType.TRACE}
-                      layout="menu"
-                    />
+                    >
+                      {({ disabled, totalCount }) => (
+                        <Button
+                          variant="ghost"
+                          size="sm"
+                          disabled={disabled !== undefined}
+                          className="w-full justify-start gap-2 font-normal"
+                        >
+                          <ListPlus className="h-4 w-4" />
+                          <span className="text-sm">Add to queue</span>
+                          {totalCount > 0 ? (
+                            <span className="bg-primary/50 text-primary-foreground ml-auto flex h-3.5 w-fit items-center justify-center rounded-sm px-1 text-xs shadow-xs">
+                              {totalCount > 99 ? "99+" : totalCount}
+                            </span>
+                          ) : null}
+                        </Button>
+                      )}
+                    </AnnotationQueueItemDropdownMenuController>
                   </>
                 )}
                 <CommentDrawerButton
@@ -213,12 +229,29 @@ export const TraceDetailViewHeader = memo(function TraceDetailViewHeader({
                   }}
                   size="sm"
                 />
-                <CreateNewAnnotationQueueItem
+                <AnnotationQueueItemDropdownMenuController
                   projectId={projectId}
                   objectId={trace.id}
                   objectType={AnnotationQueueObjectType.TRACE}
-                  size="sm"
-                />
+                >
+                  {({ disabled, totalCount }) => (
+                    <Button
+                      variant="secondary"
+                      size="sm"
+                      disabled={disabled !== undefined}
+                      className="rounded-l-none rounded-r-md border-l-2"
+                    >
+                      <span className="relative mr-1 text-xs">
+                        <ChevronDown className="h-3 w-3" />
+                        {totalCount > 0 ? (
+                          <span className="bg-primary text-primary-foreground absolute -top-1 left-2.5 flex h-3 min-w-3 items-center justify-center rounded-sm px-0.5 text-[8px] font-bold shadow-xs">
+                            {totalCount > 99 ? "99+" : totalCount}
+                          </span>
+                        ) : null}
+                      </span>
+                    </Button>
+                  )}
+                </AnnotationQueueItemDropdownMenuController>
               </div>
             )}
             <CommentDrawerButton
